+++ src/m_config.c
@@ -56,7 +56,7 @@ extern char     *packageconfig;
 extern dboolean vanilla;
 extern dboolean togglingvanilla;
 
-#define NUMCVARS                                    180
+#define NUMCVARS                                    181
 
 #define CONFIG_VARIABLE_INT(name, set)              { #name, &name, DEFAULT_INT,           set          }
 #define CONFIG_VARIABLE_INT_UNSIGNED(name, set)     { #name, &name, DEFAULT_INT_UNSIGNED,  set          }
@@ -100,6 +100,7 @@ static default_t cvars[NUMCVARS] =
     CONFIG_VARIABLE_INT          (autouse,                                           BOOLVALUEALIAS     ),
     CONFIG_VARIABLE_INT          (centerweapon,                                      BOOLVALUEALIAS     ),
     CONFIG_VARIABLE_INT          (con_backcolor,                                     NOVALUEALIAS       ),
+    CONFIG_VARIABLE_INT          (con_edgecolor,                                     NOVALUEALIAS       ),
     CONFIG_VARIABLE_INT          (con_obituaries,                                    BOOLVALUEALIAS     ),
     CONFIG_VARIABLE_INT          (con_timestamps,                                    BOOLVALUEALIAS     ),
     CONFIG_VARIABLE_INT          (crosshair,                                         CROSSHAIRVALUEALIAS),
@@ -624,6 +625,9 @@ static void M_CheckCVARs(void)
 
     if (con_backcolor < con_backcolor_min || con_backcolor > con_backcolor_max)
         con_backcolor = con_backcolor_default;
+
+    if (con_edgecolor < con_edgecolor_min || con_edgecolor > con_edgecolor_max)
+        con_edgecolor = con_edgecolor_default;
 
     if (con_obituaries != false && con_obituaries != true)
         con_obituaries = con_obituaries_default;